In the Main Activity set if statement, if saveInstanceState is null, so when I rotate it, it will not overlap. and add addToBackStack and setReorderingAllowed to save when the dice is roll.
In DiceFragment, initial the currenDice and TextView and set it to null. override the onSaveInstanceState function, and use outState.putString() to store the string key(which is const val) and value in the bundle, so always store the TextView in currentDice, if TextView is null, it behaves like TextView.text.

override the onViewCreated function, if savedInstanceState is not null, then initial the new number in order to get the currentDice. If TextView in not null, get the new Number, if it is null, store it as "0"